chore(lint): run type-aware lint and typecheck in the pre-commit hook
The hook ran plain `oxlint --fix`, so it enforced about 30 fewer rules than
`bun run lint` and could not see the breakage its own fixes caused.

- lint-staged now runs `oxlint --type-aware --fix`. The cost is about 1s:
  tsgolint builds the TypeScript program either way, so one staged file costs
  nearly what the whole repo costs.
- The hook then runs `bun run typecheck`. This is what catches an unsound
  autofix before the commit lands. Turbo caches it per package: ~0.2s when
  nothing changed, 1-2s for one package, 8s cold.

Fixes the `interface` rewrite class at the source. `consistent-type-definitions`
is context-blind — verified on oxlint 1.78, it rewrites `interface` inside
`declare module` blocks, across same-name declaration merges, and on
self-referential types, each of which only compiles as an `interface`. Both
augmentations now live in `.d.ts` files, which `.oxlintrc.json` already exempts,
so neither needs a disable comment:

- apps/web/src/router-register.d.ts     (TanStack Router `Register`)
- apps/web/src/table-column-meta.d.ts   (TanStack Table `ColumnMeta.sticky`)

Both files carry a top-level import, which is what keeps them modules and makes
the block augment the package instead of replacing it. Verified with a probe
file: `<Link to="/definitely-not-a-route">` still fails against the same route
union, and `meta?.sticky` still resolves.

AGENTS.md rule 7 records the convention.
