+++ .githooks/pre-commit
@@ -1,2 +1,13 @@
 #!/usr/bin/env sh
+set -e
+
+# Formats and lints the staged files, and applies the machine fixes.
 bunx lint-staged
+
+# `oxlint --fix` above can rewrite code in ways the linter cannot check. The clearest
+# example is `consistent-type-definitions`, which rewrites `interface` to `type` inside
+# `declare module` blocks and across declaration merges, where only `interface` is legal.
+# Typecheck is what sees this class of breakage, so it must run before the commit lands.
+# Turbo caches the result per package: about 0.2s when nothing changed, 1-2s for one
+# package, 8s for a cold cache.
+bun run typecheck
